+++ packages/sitecore-jss-react/src/index.ts
@@ -81,6 +81,7 @@ export {
   fetchFEaaSComponentServerProps,
 } from './components/FEaaSComponent';
 export { FEaaSWrapper } from './components/FEaaSWrapper';
+export { ComponentLibraryLayout } from './components/ComponentLibraryLayout';
 export {
   BYOCComponent,
   BYOCComponentParams,