+++ packages/sitecore-jss/src/editing/index.ts
@@ -8,10 +8,17 @@ export {
   handleEditorAnchors,
   Metadata,
   getJssPagesClientData,
+  addComponentUpdateHandler,
   EDITING_ALLOWED_ORIGINS,
   QUERY_PARAM_EDITING_SECRET,
   PAGES_EDITING_MARKER,
+  COMPONENT_LIBRARY_READY_MESSAGE,
+  ComponentUpdateEventArgs,
 } from './utils';
+export {
+  RestComponentLayoutService,
+  ComponentLayoutRequestParams,
+} from './rest-component-layout-service';
 export {
   DefaultEditFrameButton,
   DefaultEditFrameButtons,
@@ -23,5 +30,5 @@ export {
   EditButtonTypes,
   mapButtonToCommand,
 } from './edit-frame';
-export { RenderMetadataQueryParams } from './models';
+export { RenderMetadataQueryParams, RenderComponentQueryParams } from './models';
 export { LayoutKind, MetadataKind } from './models';