+++ packages/sitecore-jss/src/layout/rest-layout-service.ts
@@ -88,9 +88,7 @@ export class RestLayoutService extends LayoutServiceBase {
       language,
       this.serviceConfig.siteName
     );
-    const fetcher = this.serviceConfig.dataFetcherResolver
-      ? this.serviceConfig.dataFetcherResolver<LayoutServiceData>(req, res)
-      : this.getDefaultFetcher<LayoutServiceData>(req, res);
+    const fetcher = this.getFetcher(req, res);
 
     const fetchUrl = this.resolveLayoutServiceUrl('render');
 
@@ -172,12 +170,18 @@ export class RestLayoutService extends LayoutServiceBase {
     };
   };
 
+  protected getFetcher = (req?: IncomingMessage, res?: ServerResponse) => {
+    return this.serviceConfig.dataFetcherResolver
+      ? this.serviceConfig.dataFetcherResolver<LayoutServiceData>(req, res)
+      : this.getDefaultFetcher<LayoutServiceData>(req, res);
+  };
+
   /**
    * Resolves layout service url
    * @param {string} apiType which layout service API to call ('render' or 'placeholder')
    * @returns the layout service url
    */
-  protected resolveLayoutServiceUrl(apiType: 'render' | 'placeholder'): string {
+  protected resolveLayoutServiceUrl(apiType: 'render' | 'placeholder' | 'component'): string {
     const { apiHost = '', configurationName = 'jss' } = this.serviceConfig;
 
     return `${apiHost}/sitecore/api/layout/${apiType}/${configurationName}`;